+++ packages/Bio/src/utils/monomer-lib/monomer-lib.ts
@@ -19,25 +19,18 @@ import {NOTATION} from '@datagrok-libraries/bio/src/utils/macromolecule';
 import {helmTypeToPolymerType} from '@datagrok-libraries/bio/src/monomer-works/monomer-works';
 import {getUserLibSettings} from '@datagrok-libraries/bio/src/monomer-works/lib-settings';
 import {UserLibSettings} from '@datagrok-libraries/bio/src/monomer-works/types';
-import {getMonomerHandleArgs} from '@datagrok-libraries/bio/src/helm/helm-helper';
 import {GapOriginals} from '@datagrok-libraries/bio/src/utils/macromolecule/consts';
 
-import {AmbiguousWebEditorMonomer, GapWebEditorMonomer, MissingWebEditorMonomer} from './web-editor-monomer-dummy';
-import {LibraryWebEditorMonomer} from './web-editor-monomer-of-library';
+import {MonomerLibBase, MonomerLibDataType} from './monomer-lib-base';
 
 import '../../../css/cell-renderer.css';
 
 import {_package} from '../../package';
 
-const monomerRe = /[\w()]+/;
-//** Do not mess with monomer symbol with parenthesis enclosed in square brackets */
-const ambMonomerRe = RegExp(String.raw`\(${monomerRe}(,${monomerRe})+\)`);
 
 /** Wrapper for monomers obtained from different sources. For managing monomere
  * libraries, use MolfileHandler class instead */
-export class MonomerLib implements IMonomerLib {
-  private _monomers: { [polymerType: string]: { [monomerSymbol: string]: Monomer } } = {};
-  private _onChanged = new Subject<any>();
+export class MonomerLib extends MonomerLibBase implements IMonomerLib {
   private _duplicateMonomers: { [polymerType: string]: { [monomerSymbol: string]: Monomer[] } } = {};
   public get duplicateMonomers(): { [polymerType: string]: { [monomerSymbol: string]: Monomer[] } } {
     return this._duplicateMonomers;
@@ -49,11 +42,11 @@ export class MonomerLib implements IMonomerLib {
   private duplicatesNotified: boolean = false;
 
   constructor(
-    monomers: { [polymerType: string]: { [monomerSymbol: string]: Monomer } },
+    monomers: MonomerLibDataType,
     public readonly source: string | undefined = undefined,
     public readonly error: string | undefined = undefined,
   ) {
-    this._monomers = monomers;
+    super(monomers);
     for (const [_monomerType, monomersOfType] of Object.entries(this._monomers)) {
       for (const [_monomerSymbol, monomer] of Object.entries(monomersOfType))
         monomer.lib = this;
@@ -67,46 +60,6 @@ export class MonomerLib implements IMonomerLib {
         resJSON.push({...m, lib: undefined, wem: undefined});
     }
     return resJSON;
-  }
-
-  /** Creates missing {@link Monomer} */
-  addMissingMonomer(polymerType: PolymerType, monomerSymbol: string): Monomer {
-    let mSet = this._monomers[polymerType];
-    if (!mSet)
-      mSet = this._monomers[polymerType] = {};
-
-    let monomerName: string = monomerSymbol;
-    if (monomerSymbol === GapOriginals[NOTATION.HELM])
-      monomerName = 'Gap';
-    else if (polymerType === PolymerTypes.PEPTIDE && monomerSymbol === 'X')
-      monomerName = 'Any';
-    else if (polymerType === PolymerTypes.RNA && monomerSymbol === 'N')
-      monomerName = 'Any';
-
-    const m = mSet[monomerSymbol] = {
-      [REQ.SYMBOL]: monomerSymbol,
-      [REQ.NAME]: monomerName,
-      [REQ.MOLFILE]: '',
-      [REQ.AUTHOR]: 'MISSING',
-      [REQ.ID]: -1,
-      [REQ.RGROUPS]:
-        wu.count(1).take(9).map((i) => {
-          return {
-            /* eslint-disable no-multi-spaces */
-            // Samples                        //  PEPTIDE     RNA
-            [RGP.CAP_GROUP_SMILES]: '',       // '[*:1][H]'  '[*:1][H]'
-            [RGP.ALTERNATE_ID]: '',           // 'R1-H'      'R1-H'
-            [RGP.CAP_GROUP_NAME]: '',         // 'H'         'H'
-            [RGP.LABEL]: `R${i.toString()}`,  // 'R1'        'R1'
-            /* eslint-enable no-multi-spaces */
-          } as RGroup;
-        }).toArray(),
-      [REQ.SMILES]: '',
-      [REQ.POLYMER_TYPE]: polymerType,
-      [REQ.MONOMER_TYPE]: undefined as unknown as MonomerType, // TODO: Can we get monomerType from atom of POM
-      [REQ.CREATE_DATE]: null,
-    } as Monomer;
-    return m;
   }
 
   getMonomer(polymerType: PolymerType | null, argMonomerSymbol: string): Monomer | null {
@@ -192,10 +145,6 @@ export class MonomerLib implements IMonomerLib {
       return criterion;
     });
     return monomers.map((monomer) => monomer?.symbol!);
-  }
-
-  get onChanged(): Observable<any> {
-    return this._onChanged;
   }
 
   private _updateLibInt(lib: IMonomerLib): void {
@@ -335,70 +284,6 @@ export class MonomerLib implements IMonomerLib {
         ui.divText(`Monomer '${monomerSymbol}' of type '${polymerType}' not found.`),
         ui.divText('Open the Context Panel, then expand Manage Libraries'),
       ]));
-    }
-    return res;
-  }
-
-  /** Substitutes {@link org.helm.webeditor.Monomers.getMonomer()} */
-  getWebEditorMonomer(a: HelmAtom | HelmType, argName?: string): IWebEditorMonomer | null {
-    const [biotype, elem] = getMonomerHandleArgs(a, argName);
-    const pt = helmTypeToPolymerType(biotype);
-
-    /** Get or create {@link Monomer} object (in case it is missing in monomer library current config) */
-    let m: Monomer | null = this.getMonomer(pt, elem);
-    if (m && biotype == HelmTypes.LINKER && m[REQ.RGROUPS].length != 2) {
-      // Web Editor expects null
-      return null;
-    }
-    if (m && biotype == HelmTypes.SUGAR && m[REQ.RGROUPS].length != 3) {
-      // Web Editor expects null
-      return null;
-    }
-    if (!m /* && biotype != HelmTypes.LINKER*/)
-      m = this.addMissingMonomer(pt, elem);
-
-    /** Get or create {@link org,helm.WebEditorMonomer} */
-    let resWem: IWebEditorMonomer | null = m.wem ?? null;
-    if (!resWem) {
-      if (elem === '*')
-        resWem = m.wem = new GapWebEditorMonomer(biotype, elem);
-      else if (
-        (biotype === 'HELM_NUCLETIDE' && elem === 'N') ||
-        (biotype === 'HELM_AA' && elem === 'X') ||
-        (biotype === 'HELM_CHEM' && false) || // TODO: Ambiguous monomer for CHEM
-        ambMonomerRe.test(elem) // e.g. (A,R,_)
-      )
-        resWem = m.wem = new AmbiguousWebEditorMonomer(biotype, elem);
-      else if (!m.lib)
-        resWem = m.wem = new MissingWebEditorMonomer(biotype, elem);
-
-      if (!resWem)
-        resWem = m.wem = LibraryWebEditorMonomer.fromMonomer(biotype, m, this);
-    }
-
-    return resWem!;
-  }
-
-  getRS(smiles: string): { [r: string]: string } {
-    const newS = smiles.match(/(?<=\[)[^\][]*(?=])/gm);
-    const res: { [name: string]: string } = {};
-    let el = '';
-    let digit;
-    if (!!newS) {
-      for (let i = 0; i < newS.length; i++) {
-        if (newS[i] != null) {
-          if (/\d/.test(newS[i])) {
-            digit = newS[i][newS[i].length - 1];
-            newS[i] = newS[i].replace(/[0-9]/g, '');
-            for (let j = 0; j < newS[i].length; j++) {
-              if (newS[i][j] != ':')
-                el += newS[i][j];
-            }
-            res['R' + digit] = el;
-            el = '';
-          }
-        }
-      }
     }
     return res;
   }